Added an OD tab and a config tab, both need debugging
An OD tab was added that controls both the OPX and a redPitaya.

Added a folder with a simple GUI to display/modify config files for the
OPX. It is far from ready but some of the basic ingredients are there.

Efforts were made in order to render the tabs compatible to eachother.
For example, connecting the OPX in one tab should make it 'connected' in
other tabs. This still needs some debugging.